Separate code that works with task state into ITask to remove code duplication

+++ Logic/core/task/TaskParameters.h
@@ -30,7 +30,7 @@ namespace scilla {
 
 struct TaskParameters {
     uint8_t priority;
-    uint32_t stackSize;
+    uint16_t stackSize;
     void* other;
 };
